pipeline/requirements.txt
# Core
pillow>=10.2.0
numpy>=1.26.0

# Optional OCR and layout dependencies
paddleocr>=2.7.0
opencv-python>=4.9.0.80
pdf2image>=1.17.0

# Embeddings
sentence-transformers>=2.6.0

# Utilities
rapidfuzz>=3.6.2
